% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{general_operator}
\alias{general_operator}
\alias{gt}
\alias{gte}
\alias{eq}
\alias{ne}
\alias{lt}
\alias{lte}
\alias{in}
\alias{inside}
\alias{nin}
\alias{and}
\alias{not}
\alias{nor}
\alias{or}
\alias{exists}
\alias{type}
\alias{expr}
\alias{jsonSchema}
\alias{mod}
\alias{regex}
\alias{type}
\alias{text}
\alias{where}
\alias{comment}
\alias{elemMatch}
\alias{meta}
\alias{slice}
\alias{$}
\title{Mongo Query Operators}
\usage{
general_operator(name, x, allowed_queries)

gt(x)

gte(x)

eq(x)

ne(x)

lt(x)

lte(x)

"in"(x)

inside(x)

nin(x)

and(x)

not(x)

nor(x)

or(x)

exists(x)

type(x)

expr(x)

jsonSchema(x)

mod(x)

regex(x)

type(x)

text(x)

where(x)

comment(x)

elemMatch(x)

meta(x)

slice(x)

"$"(x)
}
\value{
a list that can be parsed into a part of/whole MongoDB query
}
\description{
Mongo Query Operators
}
\details{
Some query operators already have meaning in R language such as: in
for you to get around this you can refer to is as `in`(in surrounded by backtick) or inside
}
