% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paser.R
\name{build_query}
\alias{build_query}
\alias{is_empty_query}
\title{Query Parsing Functions}
\usage{
build_query(x, prettify = FALSE)

is_empty_query(x)
}
\arguments{
\item{x}{R object(list or vector) that can be parsed into ason}

\item{prettify}{flag for jsonlite::toJSON adds line breaks in output}
}
\value{
character string that is a properly formated json ready for querying the database
}
\description{
Query Parsing Functions
}
