% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_utilities.R
\name{query}
\alias{query}
\title{Query
This function is a wrapper used to make it possible to build queries using magrittr pipes}
\usage{
query(x, ...)
}
\arguments{
\item{x}{an unfinished mongo query object}

\item{...}{additional operators to be included}
}
\value{
monogo query object with additional operators included
}
\description{
Query
This function is a wrapper used to make it possible to build queries using magrittr pipes
}
