% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfastmongo.R
\docType{package}
\name{rfastmongo}
\alias{rfastmongo}
\alias{rfastmongo-pacakge}
\alias{rfastmongo-package}
\title{Build mongo queries inside R with help of R function}
\description{
Build MongoDB queries using r functions so that you don't have to worry about remebering monog's syntax, use R's autocompletion.
}
