% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_implementation.R
\name{operators}
\alias{operators}
\alias{gt}
\alias{gte}
\alias{eq}
\alias{ne}
\alias{lt}
\alias{lte}
\alias{in}
\alias{inside}
\alias{nin}
\alias{and}
\alias{not}
\alias{nor}
\alias{or}
\alias{exists}
\alias{type}
\alias{expr}
\alias{jsonSchema}
\alias{mod}
\alias{regex}
\alias{type}
\alias{text}
\alias{where}
\alias{geoIntersects}
\alias{geoWithin}
\alias{near}
\alias{nearSphere}
\alias{all}
\alias{size}
\alias{bitsAllClear}
\alias{bitsAllSet}
\alias{bitsAnyClear}
\alias{bitsAnySet}
\alias{comment}
\alias{elemMatch}
\alias{meta}
\alias{slice}
\alias{currentDate}
\alias{inc}
\alias{min}
\alias{max}
\alias{mul}
\alias{rename}
\alias{set}
\alias{setOnInsert}
\alias{unset}
\alias{addToSet}
\alias{pop}
\alias{pull}
\alias{push}
\alias{pullAll}
\alias{each}
\alias{position}
\alias{slice}
\alias{sort}
\alias{bit}
\title{Query Operators}
\usage{
gt(x)

gte(x)

eq(x)

ne(x)

lt(x)

lte(x)

"in"(x)

inside(x)

nin(x)

and(x)

not(x)

nor(x)

or(x)

exists(x)

type(x)

expr(x)

jsonSchema(x)

mod(x)

regex(x)

type(x)

text(x)

where(x)

geoIntersects(x)

geoWithin(x)

near(x)

nearSphere(x)

all(x)

size(x)

bitsAllClear(x)

bitsAllSet(x)

bitsAnyClear(x)

bitsAnySet(x)

comment(x)

elemMatch(x)

meta(x)

slice(x)

currentDate(x)

inc(x)

min(x)

max(x)

mul(x)

rename(x)

set(x)

setOnInsert(x)

unset(x)

addToSet(x)

pop(x)

pull(x)

push(x)

pullAll(x)

each(x)

position(x)

slice(x)

sort(x)

bit(x)
}
\arguments{
\item{x}{to be compared in operator}
}
\value{
a list that can be parsed into a part of/whole MongoDB query
}
\description{
Operators that can be used inside a MongoDB Query available as r function
}
\details{
Some query operators already have meaning in R language such as: in
for you to get around this you can refer to is as `in`(in surrounded by backtick) or inside
}
