% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paser.R
\name{parse_query}
\alias{parse_query}
\alias{is_empty_query}
\title{Query Parsing Functions}
\usage{
parse_query(x, prettify = FALSE, drop_nulls = TRUE)

is_empty_query(x)
}
\arguments{
\item{x}{R object(list or vector) that can be parsed into ason}

\item{prettify}{flag for jsonlite::toJSON adds line breaks in output}

\item{drop_nulls}{check if each opearator has values and remove it if it is empty}
}
\value{
character string that is a properly formated json ready for querying the database
}
\description{
Query Parsing Functions
}
