% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{find_query}
\alias{find_query}
\title{Top level object that stores all of operators that make up a query}
\usage{
find_query(...)
}
\arguments{
\item{...}{find operators}
}
\value{
nested lists that can be parsed into a MongoDB query, or JSON string
}
\description{
Top level object that stores all of operators that make up a query
}
